Fixes two small issues
1) DbEngine.Find creates a read-only transaction when starting a find,
but this is not closed if the underlying enumerator never reaches the end.
Adding a finally block around the pertitnent yield statements allows
the enumerator to be disposed, which closes the transaction.

2) If an index is created based on LiteDBDocument, it throws an error
because it's trying to add a duplicate mapping to the dictionary. This
avoids duplicates,but only taking the first one found (which should be
overriding instance.