Fixed "Property was not mapped into BsonDocument" errors on Expression queries that involve enum properties
It seems that the latest Xamarin (which recently moved to Mono 5.0+ under the hood) compiles Expressions differently - in particular references to Enum properties are now wrapped in a Convert() call.

e.g. col.Find(t => t.MyEnumProp == MyEnum.Something) 
 - used to compile to a clean "t.MyEnumProp" on the left hand side
 - now compiles to "Convert(t.MyEnumProp, Int32)" in latest Xamarin

This code change caters for the potential Convert() wrapping and extracts the inner Operand in order to restore the expected behaviour of the GetPath() method, which in turn is used in various Mapper and Query areas of LiteDb.