Fix serialization of non-generic IDictionary, for example Hashtable.
It used to fail due to ``Exception calling "Serialize" with "2" argument(s): "Index was outside the bounds of the array."``
because the original code assumed IDictionary is generic and requested invalid ``GetGenericArguments()[1]``.

Also added the test covering the problem case.

Also tweaked my previous fix of deserializing dictionaries, to avoid redundant "expensive" casts.