Fix race condition in #2536
This fixes a race condition in BsonMapper, caused by a fix to a
different issue in #2493.

It seems like the current approach of checking the dictionary twice is
deliberate. That said, I don't believe reading from a dictionary
that may be in the process of being updated is actually safe to begin
with.